init.c k_init.c xa_shel.c k_main.c k_main.h xa_global.c xa_global.h: - implement shel_write(SWM_REZCHANGE) - drop last parameter for dispatch_shutdown - change event-callback-parameter from bool to short to avoid casts

+++ xaaes/src.km/xa_global.h
@@ -40,7 +40,7 @@
 extern struct nova_data *nova_data;
 extern struct cookie_mvdi mvdi_api;
 
-extern unsigned long next_res;
+extern unsigned short next_res;
 
 extern char version[32];
 extern char vversion[128];
